refactor: use Bun.build for CSS modules instead of DIY plugin
The docs build now uses Bun's bundler natively instead of a custom CSS
modules plugin:

- scripts/docs.tsx is a thin orchestrator. It calls Bun.build() over
  scripts/docs-render.tsx, gets back the bundled JS plus a single
  emitted CSS asset (Bun handles class-name scoping per file), runs
  the bundled entry to write HTML, and copies the CSS asset to
  .build/docs/style.css.
- scripts/docs-render.tsx is what was previously the entry — same
  glob → getFileNode → nestPathNodes → writeDocs() flow.

Removes the DIY pieces this replaces:
- modules/docs/util/cssModules.ts (custom Bun plugin)
- scripts/docs-preload.ts (--preload entry that registered the plugin)
- The --preload flag in the package script

writeDocs() no longer emits the stylesheet — the orchestrator owns
that step now. README updated to describe the new pipeline. `Bun`
added to Biome globals so lint accepts Bun.build().
