--- scripts/docs-preload.ts
@@ -1,4 +1,0 @@
-// Loaded via `--preload` so the CSS-modules plugin is registered before any other module is loaded.
-import { registerCssModulesPlugin } from "../modules/docs/util/cssModules.js";
-
-registerCssModulesPlugin();
